feat: Add HeaderCell (#1)
